homeの表示処理をcontrollers/home.goに移動
